+++ app/api/v1/user/create-user/route.js
@@ -1,5 +1,5 @@
 import connectDB from "@/lib/db/configs/connection";
-import { Individual } from "@/lib/db/models/user.model";
+import DB_MODELS from "@/utils/modelsEnum";
 import { auth, currentUser } from "@clerk/nextjs";
 import { unauthorized } from '@/utils/responses';
 
@@ -10,11 +10,11 @@ export async function GET() {
 	if (!userId) {
 		return unauthorized('You must be signed in to create a user');
 	}
-	const individual = await Individual.findOne({ clerk_user_id: userId });
+	const individual = await DB_MODELS.USER.findOne({ clerk_user_id: userId });
 	
 	if (!individual) {
 		const fullName = user.firstName + (user.lastName ? " " + user.lastName : "")
-		const new_individual = new Individual({
+		const new_individual = new DB_MODELS.USER({
 			email: user.emailAddresses[0].emailAddress,
 			name: fullName,
 			clerk_user_id: user.id,
