+++ constants/CreateHabitForm/Colors/index.js
@@ -1,0 +1,89 @@
+
+const themeColors = [
+	{
+		name: 'red-variant-1',
+		color: '#fa923c',
+	},
+	{
+		name: 'orange-variant-1',
+		color: '#fe913a',
+	},
+	{
+		name: 'yellow-variant-1',
+		color: '#fabf22',
+	},
+	{
+		name: 'yellow-variant-2',
+		color: '#facc16',
+	},
+	{
+		name: 'green-variant-1',
+		color: '#a2e637',
+	},
+	{
+		name: 'green-variant-2',
+		color: '#50e07f',
+	},
+	{
+		name: 'green-variant-3',
+		color: '#36d298',
+	},
+	{
+		name: 'blue-variant-1',
+		color: '#2dd4c0',
+	},
+	{
+		name: 'blue-variant-2',
+		color: '#20d3ec',
+	},
+	{
+		name: 'blue-variant-3',
+		color: '#39bffb',
+	},
+	{
+		name: 'blue-variant-4',
+		color: '#5fa6fb',
+	},
+	{
+		name: 'purple-variant-1',
+		color: '#818df9',
+	},
+	{
+		name: 'purple-variant-2',
+		color: '#9987d9',
+	},
+	{
+		name: 'purple-variant-3',
+		color: '#c085fc',
+	},
+	{
+		name: 'pink-variant-1',
+		color: '#e978f9',
+	},
+	{
+		name: 'pink-variant-2',
+		color: '#f572b5',
+	},
+	{
+		name: 'pink-variant-3',
+		color: '#fb7387',
+	},
+	{
+		name: 'gray-variant-1',
+		color: '#94a2b9',
+	},
+	{
+		name: 'gray-variant-2',
+		color: '#9da3ae',
+	},
+	{
+		name: 'gray-variant-3',
+		color: '#a2a3a2',
+	},
+	{
+		name: 'gray-variant-4',
+		color: '#a6a29f',
+	}
+];
+
+export default themeColors
