+++ constants/CreateHabitForm/Icons/index.js
@@ -1,0 +1,111 @@
+import { CiWavePulse1 } from "react-icons/ci";
+import { IoAlarmOutline } from "react-icons/io5";
+import { CiApple } from "react-icons/ci";
+import { FaBed } from "react-icons/fa";
+import { FaWallet } from "react-icons/fa";
+import { FaHeartCircleBolt } from "react-icons/fa6";
+import { TbMoodKid } from "react-icons/tb";
+import { CgGym } from "react-icons/cg";
+import { FaBook } from "react-icons/fa";
+import { FaTerminal } from "react-icons/fa";
+import { IoColorPaletteOutline } from "react-icons/io5";
+import { FaYinYang } from "react-icons/fa";
+import { CiMusicNote1 } from "react-icons/ci";
+import { FaShower } from "react-icons/fa";
+import { TbNotes } from "react-icons/tb";
+import { RiCupLine } from "react-icons/ri";
+import { TbCurrencyDollar } from "react-icons/tb";
+import { CiHeart } from "react-icons/ci";
+import { LuLeaf } from "react-icons/lu";
+import { GiConsoleController } from "react-icons/gi";
+import { FaBicycle } from "react-icons/fa";
+
+
+const HabitFormIcons = [
+	{
+		name: "pulse",
+		icon: <CiWavePulse1 />
+	},
+	{
+		name: "alarm",
+		icon: <IoAlarmOutline />
+	},
+	{
+		name: "apple",
+		icon: <CiApple />
+	},
+	{
+		name: "bed",
+		icon: <FaBed />
+	},
+	{
+		name: "wallet",
+		icon: <FaWallet />
+	},
+	{
+		name: "heartBolt",
+		icon: <FaHeartCircleBolt />
+	},
+	{
+		name: "moodKid",
+		icon: <TbMoodKid />
+	},
+	{
+		name: "gym",
+		icon: <CgGym />
+	},
+	{
+		name: "book",
+		icon: <FaBook />
+	},
+	{
+		name: "terminal",
+		icon: <FaTerminal />
+	},
+	{
+		name: "colorPalette",
+		icon: <IoColorPaletteOutline />
+	},
+	{
+		name: "yinYang",
+		icon: <FaYinYang />
+	},
+	{
+		name: "musicNote",
+		icon: <CiMusicNote1 />
+	},
+	{
+		name: "shower",
+		icon: <FaShower />
+	},
+	{
+		name: "notes",
+		icon: <TbNotes />
+	},
+	{
+		name: "cup",
+		icon: <RiCupLine />
+	},
+	{
+		name: "dollar",
+		icon: <TbCurrencyDollar />
+	},
+	{
+		name: "heart",
+		icon: <CiHeart />
+	},
+	{
+		name: "leaf",
+		icon: <LuLeaf />
+	},
+	{
+		name: "console",
+		icon: <GiConsoleController />
+	},
+	{
+		name: "bicycle",
+		icon: <FaBicycle />
+	}
+];
+
+export default HabitFormIcons;
