+++ constants/CreateHabitForm/index.js
@@ -1,0 +1,4 @@
+import HabitFormIcons from "./Icons";
+import themeColors from "./Colors";
+
+export { HabitFormIcons, themeColors };
