+++ lib/db/models/chart.model.js
@@ -19,13 +19,13 @@ const ChartSchema = new Schema({
     type: Number,
     required: true,
   },
-  contribs: [{
+  contributions: [{
     type: Schema.Types.ObjectId,
-    ref: "Daily",
+    ref: "Contribution",
   }],
   user_id: {
     type: Schema.Types.ObjectId,
-    ref: "Teams",
+    ref: "User",
   },
   color: {
     type: String,
@@ -33,4 +33,4 @@ const ChartSchema = new Schema({
 });
 
 export const Chart =
-  mongoose.models["Chart"] || mongoose.model("Chart", ChartSchema);
+  mongoose.models["Charts"] || mongoose.model("Charts", ChartSchema);
