+++ lib/db/models/dailyContribution.model.js
@@ -13,4 +13,4 @@ const daily_schema = new Schema({
 })
 
 export const Contribution =
-  mongoose.models["Contribution"] || mongoose.model("Contribution", daily_schema);
+  mongoose.models["Contributions"] || mongoose.model("Contributions", daily_schema);
