+++ lib/db/models/user.model.js
@@ -23,9 +23,9 @@ const UserSchema = new Schema({
 	charts: [
 		{
 			type: Schema.Types.ObjectId,
-			ref: 'Charts',
+			ref: 'Chart',
 		},
 	],
 });
 
-export const User = mongoose.models['User'] || mongoose.model('User', UserSchema);
+export const User = mongoose.models['Users'] || mongoose.model('Users', UserSchema);
