+++ utils/modelsEnum.js
@@ -1,6 +1,6 @@
 import { Chart } from '@/lib/db/models/chart.model';
 import { Contribution } from '@/lib/db/models/dailyContribution.model';
-import User from '@/lib/db/models/user.model';
+import {User} from '@/lib/db/models/user.model';
 
 const DB_MODELS = {
 	USER: User,
